+++ src/CG.Logging/Extensions/ApplicationBuilderExtensions.cs
@@ -1,13 +1,10 @@
-﻿using CG.Validations;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using System;
+﻿
 
 namespace Microsoft.Extensions.Hosting
 {
     /// <summary>
     /// This class contains extension methods related to the <see cref="IApplicationBuilder"/>
-    /// types.
+    /// types, for registering types related to logging.
     /// </summary>
     public static partial class ApplicationBuilderExtensions
     {
@@ -18,23 +15,35 @@ namespace Microsoft.Extensions.Hosting
         #region Public methods
 
         /// <summary>
-        /// This method wires up any startup logic required by the "do nothing" 
-        /// logging strategy.
+        /// This method wires up any startup logic required for the logging 
+        /// services.
         /// </summary>
         /// <param name="applicationBuilder">The application builder to use 
         /// for the operation.</param>
         /// <param name="hostEnvironment">The hosting environment to use for the
         /// application.</param>
+        /// <param name="configuration">The configuration to use for the operation.</param>
         /// <returns>The value of the <paramref name="applicationBuilder"/>
         /// parameter, for chaining calls together.</returns>
         /// <exception cref="ArgumentException">This exception is thrown whenever
         /// one or more of the required parameters is missing or invalid.</exception>
-        public static IApplicationBuilder UseDoNothingLoggingStrategies(
+        public static IApplicationBuilder UseLogging(
             this IApplicationBuilder applicationBuilder,
-            IHostEnvironment hostEnvironment
+            IWebHostEnvironment hostEnvironment,
+            IConfiguration configuration
             )
         {
-            // Nothing to do here, just an example method.
+            // Validate the parameters before attempting to use them.
+            Guard.Instance().ThrowIfNull(applicationBuilder, nameof(applicationBuilder))
+                .ThrowIfNull(hostEnvironment, nameof(hostEnvironment))
+                .ThrowIfNull(configuration, nameof(configuration));
+
+            // Call the use method for the strategy.
+            applicationBuilder.UseStrategies(
+                hostEnvironment,
+                configuration,
+                "Logging"
+                );
 
             // Return the builder.
             return applicationBuilder;