+++ src/CG.Logging/GlobalUsings.cs
@@ -1,0 +1,6 @@
+﻿global using CG.Validations;
+global using Microsoft.AspNetCore.Builder;
+global using Microsoft.AspNetCore.Hosting;
+global using Microsoft.Extensions.Configuration;
+global using System;
+global using System.Diagnostics;