+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## 2022.1
 * I switched the runtime to .net 6.0
 * bug fixes
+* I added the bootstrap logger
 
 ## 2021.3